fix libevm log callback being garbage collected
